browser-server/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
playwright>=1.37.0
pydantic>=2.3.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0